api overhaul 2 #31
